requirements.txt
tensorflow
flask
numpy
opencv-python
pillow
matplotlib
scikit-learn
pandas
